changed type of vendor and product IDs to NSUInteger
git-svn-id: svn://orca.physics.unc.edu/Orca@11728 fd86f90f-2c0a-422a-b972-84106d8db31b

+++ Source/Objects/Hardware/USB/LakeShore336/ORLakeShore336Model.m
@@ -243,12 +243,12 @@ NSString* ORLakeShore336PollTimeChanged         = @"ORLakeShore336PollTimeChange
 	return [NSString stringWithFormat:@"33220 Pulser (%d)",[self tag]];
 }
 
-- (unsigned long) vendorID
+- (NSUInteger) vendorID
 {
 	return 0x1fb9;
 }
 
-- (unsigned long) productID
+- (NSUInteger) productID
 {
 	return 0x301;
 }
